finish implementation, cleanup, organize, and document
