--- babel.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  presets: [
-    [
-      "@babel/preset-env",
-      {
-        useBuiltIns: "usage",
-        corejs: 3,
-        targets: {
-          node: "current"
-        }
-      }
-    ]
-  ]
-};
