+++ package.json
@@ -5,10 +5,7 @@
   "main": "index.js",
   "scripts": {
     "jest": "jest --coverage",
-    "babel": "babel --copy-files --out-dir dist src",
-    "build": "npm run babel",
-    "prebuild": "npm run clean",
-    "clean": "rm -rf dist",
+    "build": "true",
     "fix-lint": "npm run lint -- --fix",
     "lint": "eslint src --ext .js",
     "test": "npm run jest --",
@@ -27,12 +24,6 @@
     "nock": "^10.0.6"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.3",
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "babel-eslint": "^8.0.0",
-    "babel-jest": "^24.8.0",
-    "core-js": "^3.1.4",
     "eslint": "^5.16.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-config-prettier": "^4.1.0",
