--- src/index.js
@@ -1,2 +1,0 @@
-// Useful code here hopefully
-export default () => true;
