+++ test/implementation.test.js
@@ -1,4 +1,5 @@
 const fetch = require('node-fetch');
+// import fetch from 'node-fetch';
 
 const { expect } = global;
