--- test/index.test.js
@@ -1,9 +1,0 @@
-import { assert } from 'chai';
-
-import index from '../src';
-
-describe('index', () => {
-  it('should be useful?', () => {
-    assert.isTrue(index());
-  });
-});
